feat(unicorn): unicorn api is a part of cocaine now
