--- unicorn/include/cocaine/unicorn/path.hpp
@@ -1,22 +1,0 @@
-/*
-* 2015+ Copyright (c) Anton Matveenko <antmat@yandex-team.ru>
-* All rights reserved.
-*
-* This program is free software; you can redistribute it and/or modify
-* it under the terms of the GNU General Public License as published by
-* the Free Software Foundation; either version 2 of the License, or
-* (at your option) any later version.
-*
-* This program is distributed in the hope that it will be useful,
-* but WITHOUT ANY WARRANTY; without even the implied warranty of
-* MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
-* GNU General Public License for more details.
-*/
-
-#pragma once
-
-#include <string>
-
-namespace cocaine { namespace unicorn {
-typedef std::string path_t;
-}}
