+++ unicorn/src/module.cpp
@@ -17,7 +17,7 @@
 
 #include "cocaine/detail/unicorn/zookeeper.hpp"
 
-#include "cocaine/unicorn/errors.hpp"
+#include "cocaine/detail/zookeeper/errors.hpp"
 
 #include "cocaine/service/unicorn.hpp"
 
@@ -35,7 +35,6 @@ initialize(api::repository_t& repository) {
     repository.insert<service::unicorn_service_t>("unicorn");
     repository.insert<cluster::unicorn_cluster_t>("unicorn");
     repository.insert<unicorn::zookeeper_t>("zookeeper");
-    error::registrar::add(error::unicorn_category());
     error::registrar::add(error::zookeeper_category());
 }
