--- unicorn/src/unicorn/value.cpp
@@ -1,26 +1,0 @@
-/*
-* 2015+ Copyright (c) Anton Matveenko <antmat@yandex-team.ru>
-* All rights reserved.
-*
-* This program is free software; you can redistribute it and/or modify
-* it under the terms of the GNU General Public License as published by
-* the Free Software Foundation; either version 2 of the License, or
-* (at your option) any later version.
-*
-* This program is distributed in the hope that it will be useful,
-* but WITHOUT ANY WARRANTY; without even the implied warranty of
-* MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
-* GNU General Public License for more details.
-*/
-
-#include "cocaine/unicorn/value.hpp"
-#include "cocaine/unicorn/errors.hpp"
-
-namespace cocaine { namespace unicorn {
-
-versioned_value_t::versioned_value_t(value_t _value, version_t _version) :
-    value(std::move(_value)),
-    version(std::move(_version))
-{}
-
-}}
